move confetti into separate component
